refactor: Move DependencyHelpers.ts back to src/ui
This keeps it consistent with the location of its tests,
and more importantly, will make the diffs easier to see when
the PR is accepted and squashed.
